Changed VS2005 compiler warning level to 4 (maximum number of warnings) and fixed all easy and low risk warnings.  There are 11 remaining warnings.
git-svn-id: https://216.105.38.17/p/tinycad/code/TinyCAD/trunk@568 41062d70-b26d-460a-9302-f438a8930db8

+++ src/DrawText.cpp
@@ -69,7 +69,7 @@ void CDrawText::CalcLayout()
 
 }
 
-int CDrawText::getMenuID()
+UINT CDrawText::getMenuID()
 {
 	switch (xtype)
 	{
@@ -85,7 +85,7 @@ int CDrawText::getMenuID()
 			return IDM_TOOLBUSNAME;
 	}
 
-	return -1;
+	return static_cast<UINT>(-1);
 }
 
 const TCHAR* CDrawText::GetXMLTag(ObjType t)
