+++ src/Object.h
@@ -122,7 +122,7 @@ public:
 	virtual void NewOptions();
 	virtual CString GetName() const;
 	virtual void TagResources();
-	virtual int getMenuID();
+	virtual UINT getMenuID();
 
 	virtual void LButtonUp(CDPoint, CDPoint); // The user has released the left hand button
 	virtual BOOL CanControlAspect(); // If the user holds down ctrl, can should we fix the aspect ratio?
@@ -196,7 +196,7 @@ public:
 	virtual void NewOptions();
 	virtual CString GetName() const;
 	virtual void TagResources();
-	virtual int getMenuID();
+	virtual UINT getMenuID();
 	virtual BOOL CanEdit()
 	{
 		return TRUE;
@@ -239,7 +239,7 @@ public:
 	virtual void LButtonDown(CDPoint, CDPoint);
 	virtual ObjType GetType(); // Get this object's type
 	virtual CString GetName() const;
-	virtual int getMenuID()
+	virtual UINT getMenuID()
 	{
 		return IDM_TOOLJUNC;
 	}
@@ -268,7 +268,7 @@ public:
 	virtual void LButtonDown(CDPoint, CDPoint);
 	virtual ObjType GetType(); // Get this object's type
 	virtual CString GetName() const;
-	virtual int getMenuID();
+	virtual UINT getMenuID();
 	virtual void Display(BOOL erase = TRUE);
 
 	// Extract the netlist/active points from this object
@@ -300,7 +300,7 @@ public:
 	virtual void LButtonDown(CDPoint, CDPoint);
 	virtual ObjType GetType(); // Get this object's type
 	virtual CString GetName() const;
-	virtual int getMenuID()
+	virtual UINT getMenuID()
 	{
 		return IDM_TOOLCONNECT;
 	}
@@ -338,7 +338,7 @@ public:
 	virtual void LButtonDown(CDPoint, CDPoint);
 	virtual ObjType GetType(); // Get this object's type
 	virtual CString GetName() const;
-	virtual int getMenuID()
+	virtual UINT getMenuID()
 	{
 		return IDM_TOOLORIGIN;
 	}
@@ -410,7 +410,7 @@ public:
 	void NewFont(LOGFONT *); // Change the font
 	virtual void NewOptions();
 
-	virtual int getMenuID();
+	virtual UINT getMenuID();
 	virtual void MoveField(int w, CDPoint r);
 
 	// These are used for the construction of this object
@@ -511,7 +511,7 @@ public:
 	void NewFont(LOGFONT *); // Change the font
 	virtual void NewOptions();
 
-	virtual int getMenuID();
+	virtual UINT getMenuID();
 	virtual void MoveField(int w, CDPoint r);
 
 	// These are used for the construction of this object
@@ -570,7 +570,7 @@ public:
 	}
 	CString Find(const TCHAR *);
 
-	virtual int getMenuID()
+	virtual UINT getMenuID()
 	{
 		return IDM_TOOLPOWER;
 	}
@@ -670,7 +670,7 @@ public:
 	
 	static const TCHAR *GetElectricalTypeName(int etype);
 
-	virtual int getMenuID()
+	virtual UINT getMenuID()
 	{
 		return IDM_LIBPIN;
 	}
@@ -720,7 +720,7 @@ public:
 	CDrawMetaFile(CTinyCadDoc *pDesign);
 	virtual ~CDrawMetaFile();
 
-	virtual int getMenuID()
+	virtual UINT getMenuID()
 	{
 		return IDM_TOOLSQUARE;
 	}
@@ -773,7 +773,7 @@ public:
 	// This is used for the construction of this object
 	CDrawSquare(CTinyCadDoc *pDesign, ObjType type);
 
-	virtual int getMenuID() 
+	virtual UINT getMenuID() 
 	{
 		return IsSquare() ? IDM_TOOLSQUARE : IDM_TOOLCIRCLE;
 	}
@@ -841,7 +841,7 @@ public:
 	virtual ~CDrawNoteText()
 	{
 	}
-	virtual int getMenuID() 
+	virtual UINT getMenuID() 
 	{
 		return IDM_TOOLNOTETEXT;
 	}
@@ -944,7 +944,7 @@ public:
 	virtual void Move(CDPoint, CDPoint no_snap_p);
 	virtual CDrawingObject* Store();
 	virtual CString GetName() const;
-	virtual int getMenuID();
+	virtual UINT getMenuID();
 	virtual int SetCursorEdit(CDPoint p);
 
 	virtual BOOL IsConstruction();
@@ -969,7 +969,7 @@ public:
 	virtual void Move(CDPoint, CDPoint no_snap_p);
 
 	virtual int SetCursor(CDPoint p);
-	virtual int getMenuID()
+	virtual UINT getMenuID()
 	{
 		return IDM_EDITMOVE;
 	}
@@ -990,7 +990,7 @@ public:
 	virtual void ChangeDir(int);
 
 	virtual int SetCursor(CDPoint p);
-	virtual int getMenuID()
+	virtual UINT getMenuID()
 	{
 		return IDM_EDITROTATE;
 	}
@@ -1010,7 +1010,7 @@ public:
 	virtual void Move(CDPoint, CDPoint no_snap_p);
 
 	virtual int SetCursor(CDPoint p);
-	virtual int getMenuID()
+	virtual UINT getMenuID()
 	{
 		return IDM_EDITDUP;
 	}
@@ -1030,7 +1030,7 @@ public:
 	virtual BOOL RButtonDown(CDPoint, CDPoint);
 	virtual void Move(CDPoint, CDPoint no_snap_p);
 	virtual int SetCursor(CDPoint p);
-	virtual int getMenuID()
+	virtual UINT getMenuID()
 	{
 		return IDM_EDITDRAG;
 	}
@@ -1047,7 +1047,7 @@ public:
 	virtual void Move(CDPoint, CDPoint no_snap_p);
 	virtual void Import();
 
-	virtual int getMenuID()
+	virtual UINT getMenuID()
 	{
 		return IDM_FILEIMPORT;
 	}
@@ -1093,7 +1093,7 @@ public:
 	virtual void ContextMenu(CDPoint p, UINT id);
 	virtual void ChangeDir(int NewDir);
 
-	virtual int getMenuID()
+	virtual UINT getMenuID()
 	{
 		return IDM_EDITEDIT;
 	}
@@ -1289,7 +1289,7 @@ public:
 	{
 	}
 
-	virtual int getMenuID()
+	virtual UINT getMenuID()
 	{
 		return IDM_TOOLGET;
 	}
@@ -1315,7 +1315,7 @@ public:
 	{
 		return 9;
 	}
-	virtual int getMenuID()
+	virtual UINT getMenuID()
 	{
 		return IDM_VIEWCENTRE;
 	}
@@ -1343,9 +1343,9 @@ public:
 	{
 		return 10;
 	}
-	virtual int getMenuID()
+	virtual UINT getMenuID()
 	{
-		return -1;
+		return static_cast<UINT>(-1);
 	}
 };
